[minor] rename files
