+++ sequence/kernel/quantum_manager/__init__.py
@@ -1,7 +1,7 @@
 from .base import QuantumManager
-from .ket import QuantumManagerKet
-from .density import QuantumManagerDensity
-from .fock import QuantumManagerDensityFock
+from .ket_vector import QuantumManagerKet
+from .density_matrix import QuantumManagerDensity
+from .fock_density_matrix import QuantumManagerDensityFock
 from .bell_diagonal import QuantumManagerBellDiagonal
 
 __all__ = [
